finally fix bug

+++ lab_02/include/algorithms.h
@@ -5,7 +5,7 @@
 #include "constants.h"
 
 enum ERRORS_T data_check(const data_t *const data);
-double interpolation(double **const points, const int size, const int deg, const int arg, 
+double interpolation(double **const points, const int size, const int deg, const double arg, 
         const int mode);
 double multi_interpolation(data_t *const data);
